predavanja/experiments/pred-06/pyproject.toml
[tool.poetry]
name = "visualize"
version = "0.1.0"
description = ""
authors = ["Petar Marić <petar.maric@gmail.com>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "~=8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
